+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Dropped 1.9-1.10 support
 * /vda reload now force players to quit to prevent problems
 * Now shop will be successfully registered when arena is freshly created
+* Added villagedefense.command.override permission to be able to use all game commands while being in VillageDefense game
 
 ### 3.11.2 Release (03.09.2018)
 * Temporarily merged PLCore to fix issues when using my other plugins